Change location of function
